requirements.txt
pushbots
unidecode
BeautifulSoup
goose
newspaper
haystack
celery
django=1.8.4
mysqlclient=1.3.6